app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.10.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
pytest==8.3.3
pytest-anyio==0.0.0
python-dotenv==1.0.1
nanoid==2.0.0
